# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/staff/context": {
      "get": {
        "operationId": "StaffConsoleController_context",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffContextDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Who am I, which business, which campaign is live",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/today": {
      "get": {
        "operationId": "StaffConsoleController_today",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Today's counter numbers (managers also get team breakdown)",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/password": {
      "post": {
        "operationId": "StaffConsoleController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change own password (requires current password)",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/customers/search": {
      "get": {
        "operationId": "StaffConsoleController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Phone digits, customer code or name. Empty = recently stamped.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipListItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Find a customer by phone, code or name (own business only)",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/enroll": {
      "post": {
        "operationId": "StaffConsoleController_enroll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrol a customer at the counter by phone or email (no OTP needed)",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/stamps": {
      "post": {
        "operationId": "StaffConsoleController_addStamp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStampDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStampResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add one stamp to a customer card",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/stamps/undo": {
      "post": {
        "operationId": "StaffConsoleController_undoStamp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UndoStampDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoStampResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Undo the latest stamp on a card (own within 60 s; managers any within 15 min)",
        "tags": [
          "Staff console"
        ]
      }
    },
    "/v1/staff/redemptions/redeem": {
      "post": {
        "operationId": "StaffConsoleController_redeem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a reward voucher as handed over (by id or code)",
        "tags": [
          "Staff console"
        ]
      }
    },
...
  },
  "info": {
    "title": "Loyalty Platform API",
    "description": "Multi-tenant loyalty platform — Phase 1. Bearer JWT auth; see /docs for the interactive UI.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BusinessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm5xyzabc0000abcd1234efgh"
          },
          "name": {
            "type": "string",
            "example": "Brew & Bean Coffee"
          },
          "slug": {
            "type": "string",
            "example": "brew-and-bean"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true,
            "example": "http://localhost:4000/uploads/logos/x.png"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "googleReviewUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://g.page/r/CaBcDeFgHiJkL/review"
          },
          "joinUrl": {
            "type": "string",
            "example": "http://localhost:3000/join/brew-and-bean",
            "description": "Customer registration link encoded in the QR code"
          },
          "brandColor": {
            "type": "string",
            "nullable": true,
            "example": "#4F46E5"
          },
          "stampIcon": {
            "type": "string",
            "nullable": true,
            "example": "☕",
            "description": "Default stamp emoji for cards"
          },
          "rewardIcon": {
            "type": "string",
            "nullable": true,
            "example": "🎁",
            "description": "Default reward emoji for cards"
          },
          "cardImageUrl": {
            "type": "string",
            "nullable": true,
            "description": "Default card background image URL"
          },
          "cardImageTint": {
            "type": "boolean",
            "example": true,
            "description": "Tint the card image with the colour, or show image + scrim only"
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "consentText": {
            "type": "string",
            "nullable": true
          },
          "notifyDailySummary": {
            "type": "boolean"
          },
          "notifyWeeklyDigest": {
            "type": "boolean"
          },
          "notifyStaffInactive": {
            "type": "boolean"
          },
          "suspended": {
            "type": "boolean",
            "description": "True while a platform admin has suspended the account"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "logoUrl",
          "address",
          "phone",
          "googleReviewUrl",
          "joinUrl",
          "brandColor",
          "stampIcon",
          "rewardIcon",
          "cardImageUrl",
          "cardImageTint",
          "category",
          "timezone",
          "consentText",
          "notifyDailySummary",
          "notifyWeeklyDigest",
          "notifyStaffInactive",
          "suspended",
          "createdAt"
        ]
      },
...
      "RedeemRequestDto": {
        "type": "object",
        "properties": {
          "redemptionId": {
            "type": "string",
            "description": "Redemption id from a card/search payload"
          },
          "code": {
            "type": "string",
            "example": "RX7K-9QZ2",
            "description": "Voucher code as shown on the customer card"
          }
        }
      },
...
      "CustomerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "+919876501101"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "ami@example.com"
          },
          "contact": {
            "type": "string",
            "example": "+91 98765 01101"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "contact"
        ]
      },
      "RedemptionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Redemption (voucher) id"
          },
          "code": {
            "type": "string",
            "example": "RX7K9QZ2"
          },
          "formattedCode": {
            "type": "string",
            "example": "RX7K-9QZ2"
          },
          "rewardText": {
            "type": "string",
            "example": "1 free coffee of your choice",
            "description": "Snapshotted at earn time"
          },
          "earnedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "rewardText",
          "earnedAt"
        ]
      },
      "MembershipListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Membership id"
          },
          "code": {
            "type": "string",
            "example": "7F3K9QZP"
          },
          "formattedCode": {
            "type": "string",
            "example": "7F3K-9QZP"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Private merchant notes"
          },
          "tags": {
            "example": [
              "regular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "blockedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "blockedReason": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "stampCount": {
            "type": "number",
            "example": 4
          },
          "stampsRequired": {
            "type": "number",
            "example": 10
          },
          "completedCount": {
            "type": "number",
            "example": 1
          },
          "totalStamps": {
            "type": "number",
            "example": 14
          },
          "lastStampAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingRewards": {
            "description": "Rewards waiting to be handed over",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedemptionSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "notes",
          "tags",
          "blockedAt",
          "blockedReason",
          "customer",
          "stampCount",
          "stampsRequired",
          "completedCount",
          "totalStamps",
          "lastStampAt",
          "joinedAt",
          "pendingRewards"
        ]
      },
      "CardCampaignDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Coffee Lovers Card"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stampsRequired": {
            "type": "number",
            "example": 10
          },
          "reward": {
            "type": "string",
            "example": "1 free coffee of your choice"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "PAUSED",
              "ARCHIVED"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "stampsRequired",
          "reward",
          "status"
        ]
      },
...
      "StampDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "delta": {
            "type": "number",
            "example": 1,
            "description": "Signed movement: −1 rows are undo reversals or removals"
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "description": "Present on adjustments and undos"
          },
          "completedCard": {
            "type": "boolean",
            "description": "True when this stamp completed a card and earned the reward"
          },
          "issuerType": {
            "type": "string",
            "enum": [
              "STAFF",
              "MERCHANT",
              "ADJUSTMENT"
            ]
          },
          "issuerName": {
            "type": "string",
            "example": "Ravi Kumar",
            "description": "Staff name, or \"Owner\" for merchant-issued stamps"
          }
        },
        "required": [
          "id",
          "createdAt",
          "delta",
          "reason",
          "completedCard",
          "issuerType",
          "issuerName"
        ]
      },
      "AddStampResultDto": {
        "type": "object",
        "properties": {
          "card": {
            "description": "Card state after the stamp",
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipListItemDto"
              }
            ]
          },
          "rewardEarned": {
            "type": "boolean",
            "description": "True when this stamp completed the card"
          },
          "reward": {
            "type": "string",
            "example": "1 free coffee of your choice"
          },
          "stamp": {
            "$ref": "#/components/schemas/StampDto"
          },
          "redemption": {
            "nullable": true,
            "description": "Voucher minted when this stamp completed the card",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RedemptionSummaryDto"
              }
            ]
          }
        },
        "required": [
          "card",
          "rewardEarned",
          "reward",
          "stamp",
          "redemption"
        ]
      },
...
      "RedeemResultDto": {
        "type": "object",
        "properties": {
          "redemption": {
            "$ref": "#/components/schemas/RedemptionRowDto"
          },
          "card": {
            "description": "Card state after redeeming",
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipListItemDto"
              }
            ]
          }
        },
        "required": [
          "redemption",
          "card"
        ]
      },
...
      "StaffSelfDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "STAFF",
              "MANAGER"
            ],
            "description": "MANAGER unlocks team stats and longer undo"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role"
        ]
      },
      "StaffContextDto": {
        "type": "object",
        "properties": {
          "staff": {
            "$ref": "#/components/schemas/StaffSelfDto"
          },
          "business": {
            "$ref": "#/components/schemas/BusinessDto"
          },
          "campaign": {
            "nullable": true,
            "description": "Campaign stamps are added to (null when none active)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CardCampaignDto"
              }
            ]
          }
        },
        "required": [
          "staff",
          "business",
          "campaign"
        ]
      },
      "TodayMineDto": {
        "type": "object",
        "properties": {
          "stamps": {
            "type": "number",
            "example": 34,
            "description": "Net stamps issued today (undos subtract)"
          },
          "redemptions": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "stamps",
          "redemptions"
        ]
      },
      "TodayTotalsDto": {
        "type": "object",
        "properties": {
          "stamps": {
            "type": "number",
            "example": 61
          },
          "newCustomers": {
            "type": "number",
            "example": 5
          },
          "rewardsRedeemed": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "stamps",
          "newCustomers",
          "rewardsRedeemed"
        ]
      },
      "TodayTeamMemberDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Meera Iyer"
          },
          "stamps": {
            "type": "number",
            "example": 27
          },
          "redemptions": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "id",
          "name",
          "stamps",
          "redemptions"
        ]
      },
      "TodayDto": {
        "type": "object",
        "properties": {
          "mine": {
            "$ref": "#/components/schemas/TodayMineDto"
          },
          "totals": {
            "nullable": true,
            "description": "Managers only",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TodayTotalsDto"
              }
            ]
          },
          "team": {
            "nullable": true,
            "description": "Managers only",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayTeamMemberDto"
            }
          }
        },
        "required": [
          "mine",
          "totals",
          "team"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password for verification"
          },
          "newPassword": {
            "type": "string",
            "description": "New password (min 8 chars)"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "EnrollDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "+91 98765 43210",
            "description": "The customer's phone number or email address — either identifies them"
          },
          "name": {
            "type": "string",
            "example": "Asha Patel"
          },
          "marketingConsent": {
            "type": "boolean",
            "description": "Only send true when the customer explicitly agreed at the counter."
          }
        },
        "required": [
          "identifier"
        ]
      },
      "EnrollResultDto": {
        "type": "object",
        "properties": {
          "card": {
            "description": "The (new or existing) card",
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipListItemDto"
              }
            ]
          },
          "alreadyMember": {
            "type": "boolean",
            "description": "True when this phone already held a card here"
          },
          "isNewCustomer": {
            "type": "boolean",
            "description": "True when no customer account existed for this phone yet"
          }
        },
        "required": [
          "card",
          "alreadyMember",
          "isNewCustomer"
        ]
      },
      "AddStampDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "description": "Membership id from a search result"
          }
        },
        "required": [
          "membershipId"
        ]
      },
      "UndoStampDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string",
            "description": "Membership id whose latest stamp should be taken back"
          }
        },
        "required": [
          "membershipId"
        ]
      },
      "UndoStampResultDto": {
        "type": "object",
        "properties": {
          "card": {
            "description": "Card state after the undo",
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipListItemDto"
              }
            ]
          },
          "voucherVoided": {
            "type": "boolean",
            "description": "True when the undone stamp had completed the card and its voucher was voided"
          }
        },
        "required": [
          "card",
          "voucherVoided"
        ]
      },
...
    }
  }
}